+++ Engine.Tests.Unit/StubRng.h
@@ -5,14 +5,14 @@ namespace engine {
 	class StubRng : public IRng {
 
 	public:
-		StubRng::StubRng(int values[]);
+		StubRng::StubRng(long long values[]);
 		virtual StubRng::~StubRng() override;
 
-		int NextInteger() override;
+		long long NextInteger() override;
 
 	private:
 		int _pointer;
-		int* _values;
+		long long* _values;
 
 	};
 }
